requirements.txt
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
pypdf
sentence-transformers
email-validator
python-multipart
streamlit

-e .